27-09-2023

+++ app/Http/Controllers/UserController.php
@@ -8,35 +8,41 @@ use Spatie\Permission\Models\Role;
 use Illuminate\Support\Facades\Hash;
 use App\Models\Cargos;
 use App\Models\Supers;
+use App\Models\Userequipos;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\DB;
 
 class UserController extends Controller
 {
+    function __construct()
+    {
+        $this->middleware('permission:listar-usuarios|crear-usuarios|editar-usuarios|eliminar-usuarios', ['only' => ['index', 'show']]);
+        $this->middleware('permission:crear-usuarios', ['only' => ['create', 'store']]);
+        $this->middleware('permission:editar-usuarios', ['only' => ['edit', 'update']]);
+        $this->middleware('permission:eliminar-usuarios', ['only' => ['destroy']]);
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $usuarios = User::leftjoin('model_has_roles','users.id','=','model_has_roles.model_id')
-                        ->leftjoin('roles','model_has_roles.role_id','=','roles.id')
-                        ->select('users.*','roles.name as rol')
-                        ->orderby('users.id')
-                        ->get();
-        //dd($usuarios);       
+        $usuarios = User::leftjoin('model_has_roles', 'users.id', '=', 'model_has_roles.model_id')
+            ->leftjoin('roles', 'model_has_roles.role_id', '=', 'roles.id')
+            ->select('users.*', 'roles.name as rol')
+            ->orderby('users.id')
+            ->get();
         return view('usuarios.index', compact('usuarios'));
     }
-
     /**
      * Show the form for creating a new resource.
      */
     public function create()
     {
         $roles = Role::pluck('name', 'name')->all();
-        $cargos = Cargos::orderBy('descripcion','ASC')->get();  
+        $cargos = Cargos::orderBy('descripcion', 'ASC')->get();
         $supers = Supers::all();
-        return view('usuarios.create', compact('roles','cargos','supers'));
+        return view('usuarios.create', compact('roles', 'cargos', 'supers'));
     }
 
     /**
@@ -47,32 +53,31 @@ class UserController extends Controller
         $this->validate($request, [
             'primerNombre' => 'required|string|max:255',
             'segundoNombre',
-            'primerApellido'=> 'required|string|max:255',
+            'primerApellido' => 'required|string|max:255',
             'segundoApellido',
             'email' => 'required|email|unique:users,email',
             'password' => 'required|same:confirm-password',
-            'sexo'=> 'required',
+            'sexo' => 'required',
             'roles' => 'required',
-            'idCargo'=> 'required'
+            'idCargo' => 'required'
         ]);
 
         $input = $request->all();
-        $input['password'] = Hash::make($input['password']);      
+        $input['password'] = Hash::make($input['password']);
 
         $user = User::create($input);
 
 
-        if (!($input['responsable']))
-        {
+        if (!($input['responsable'])) {
             $superuser = Supers::find($input['responsable']);
-            $superuser->responsable = $user->id;  
-            $superuser->save();        
+            $superuser->responsable = $user->id;
+            $superuser->save();
         }
 
         $user->assignRole($request->input('roles'));
 
         return redirect()->route('usuarios.index')
-            ->with('success', 'User created successfully');
+            ->with('success', 'Usuario creado exitosamente');
     }
 
     /**
@@ -81,7 +86,10 @@ class UserController extends Controller
     public function show(string $id)
     {
         $usuario = User::find($id);
-        return view('usuarios.show',compact('usuario'));
+        $userequipo = $usuario->equipos;
+        //dd($userequipo);
+
+        return view('usuarios.show', compact('usuario','userequipo'));
     }
 
     /**
@@ -91,46 +99,41 @@ class UserController extends Controller
     {
         $usuario = User::find($id);
         $roles = Role::pluck('name', 'name')->all();
-        $userroles = $usuario->roles->pluck('name')->all();  
-       
-        $cargos = Cargos::orderBy('descripcion','ASC')->get(); 
-        $responsable = Supers::where('responsable',$id)->first();
-        //dd($userroles);
+        $userroles = $usuario->roles->pluck('name')->all();
+        $cargos = Cargos::orderBy('descripcion', 'ASC')->get();
+        $responsable = Supers::where('responsable', $id)->first();
         $supers = Supers::all();
-        return view('usuarios.edit', compact('roles','cargos','supers','usuario','userroles','responsable'));
+        return view('usuarios.edit', compact('roles', 'cargos', 'supers', 'usuario', 'userroles', 'responsable'));
     }
-
     /**
      * Update the specified resource in storage.
      */
     public function update(Request $request, string $id)
     {
-   
+
         $this->validate($request, [
             'primerNombre' => 'required|string|max:255',
             'segundoNombre',
-            'primerApellido'=> 'required|string|max:255',
+            'primerApellido' => 'required|string|max:255',
             'segundoApellido',
-            'email' => 'required|email|unique:users,email,'.$id,            
+            'email' => 'required|email|unique:users,email,' . $id,
             'sexo',
             'roles' => 'required',
-            'idCargo'=> 'required'
+            'idCargo' => 'required'
         ]);
-        $input = $request->all(); 
-        $user = User::find($id);       
-        if (($input['responsable'])!=NULL) 
-        {
+        $input = $request->all();
+        $user = User::find($id);
+        if (($input['responsable']) != NULL) {
             $superuser = Supers::find($input['responsable']);
-            $superuser->responsable = $user->id;  
-            $superuser->save(); 
-        }   
+            $superuser->responsable = $user->id;
+            $superuser->save();
+        }
         $user->update($input);
-        
-        DB::table('model_has_roles')->where('model_id',$id)->delete();
+
+        DB::table('model_has_roles')->where('model_id', $id)->delete();
         $user->assignRole($request->input('roles'));
         return redirect()->route('usuarios.index')
-                        ->with('success','Usuario actualizado con exito!!');
-
+            ->with('success', 'Usuario actualizado con exito!!');
     }
 
     /**
@@ -138,15 +141,18 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
-        $responsable = Supers::where('responsable',$id)->first();
-        //dd($responsable);
-        if($responsable){
-            return redirect()->route('usuarios.index')
-            ->with('success', 'No se puede eliminar es responsable de area');
+        $responsable = Supers::where('responsable', $id)->first();
+        if ($responsable) {
+            return response()->json(['success' => 'El usuario No se puede eliminar es responsable de area']);
         }
-        User::find($id)->delete();     
+        $tieneEquipo = Userequipos::where('idUser', $id)->first();
 
-        return redirect()->route('usuarios.index')
-             ->with('success', 'User delete successfully');
+        if ($tieneEquipo) {
+            return response()->json(['success' => 'El usuario No se puede eliminar tiene equipos asignados']);
+        }
+        User::find($id)->delete();
+        return response()->json(['success' => 'Usuario eliminado correctamente']);
+
+
     }
 }
